lexer: Introduce remaining keywords from reference
This commit adds the remaining keywords to the tokenizer. The internals
have been changed to capture all Keywords within an umbrella enum. This
has been done to speed up checking if a token is a keyword. Prior to
this it might have involved a multi-way match. With this change it
should be as simple as a single match statement looking at the outermost
discriminant :')

Also included are:

* Small formatting changes.
* Rename "BuiltinOperator" to "Op"